tests/chaos/workloads: split out workload base class
