+++ tests/rptest/services/chaos/workloads/all.py
@@ -7,4 +7,5 @@
 # the Business Source License, use of this software will be governed
 # by the Apache License, Version 2.0
 
+from .service_base import WorkloadServiceBase
 from .list_offsets.service import ListOffsetsWorkload
